test(core): migrate UpdateMruTests (Wave C)
21 tests, 70 failure sites preserved 1:1. The three compound `guard let ... else { throw }` chained-expression patterns convert to Issue.record + return. The now-unused makeMruModel helper drops out of tests/TestHarness.swift.
