+++ tests/TestHarness.swift
@@ -19,7 +19,6 @@ struct TestRunner {
         todoPopoverStateTests()
         switcherEventTests()
         jumpEventTests()
-        updateMruTests()
         updateJumpTests()
         print("\n\(total - failures)/\(total) passed")
         if failures > 0 { exit(1) }
@@ -97,16 +96,3 @@ func paneSnapshot(_ id: String, in snapshot: AppModelSnapshot) -> PaneSnapshot? 
     allPaneSnapshots(snapshot).first { $0.id == id }
 }
 
-/// Build a model with N tabs in one group; returns the tab ids in display order.
-/// Used by MRU/cycle tests in tests/UpdateMruTests.swift until its migration.
-func makeMruModel(tabCount: Int) -> (model: AppModel, tabIds: [TabId]) {
-    var model = makeModel()
-    var ids: [TabId] = []
-    for _ in 0..<tabCount {
-        let paneId = PaneId()
-        let tabId = TabId()
-        ids.append(tabId)
-        model.groups[0].tabs.append(TabModel(id: tabId, focusedPaneId: paneId, rootNode: .leaf(PaneModel(id: paneId))))
-    }
-    return (model, ids)
-}
